Android 12 Widgets - Introduce layout previews
https://developer.android.com/about/versions/12/features/widgets#add-scalable-widget-previews

Enable a more accurate version of the widget preview by using layouts directly. This creates a much nicer animation when user drags and drops the widget in the home screen. The layout preview adapts to theme and sizes

In order to accomplish it we need to create preview layouts and attrs to define different values for each preview item.

Given the duplication in contacts_widget_item.xml the commit extracts each item into it's own layout. Future commits will fix this duplication.